64bit vector/bytevector support and cleanups
This allows vectors up to the size of the virtual address space and
bytevectors of 2^63-1 bytes on 64bit or the size the address space on
32bit.

- Introduce type aliases for our various length types and add
  ::maximumLength() for strings, symbols, vectors and bytevectors. These
  are based on the underlying storage types for the values.

- Introduce RecordLikeCell::RecordClassIdType to abstract away the types
  of record class IDs

- Rename SharedByteArray::refcount_t to SharedByteArray::RefCountType
  for consistency

- Add assertLengthValid() which is analogous to assertIndexValid
  except for validating the length of new objects

- Add adjustSlice() to factor out the slice adjustment logic from
  strings, vectors and bytevectors

- Change llcore_vector_elements_alloc to llcore_vector_alloc which will
  allocate both the vector cell and elements at once. This is required
  to properly handle exceptions when the length is loo targe and in the
  future when out-of-memory errors are supported

- Update cellTypes.celldef and the various stdlib functions to use 64bit
  lengths for vectors and bytevectors

+++ compiler/src/main/scala/planner/reportproc/BytevectorProcPlanner.scala
@@ -20,10 +20,10 @@ object BytevectorProcPlanner extends ReportProcPlanner {
         bytevectorValue.toTempValue(vt.BytevectorType)
       }
 
-      val resultTemp = ps.Temp(vt.UInt32)
+      val resultTemp = ps.Temp(vt.Int64)
       plan.steps += ps.LoadBytevectorLength(resultTemp, bytevectorTemp)
 
-      Some(TempValueToIntermediate(vt.UInt32, resultTemp)(plan.config))
+      Some(TempValueToIntermediate(vt.Int64, resultTemp)(plan.config))
 
     case _ =>
       None
